fix(olm): Use the next key id from the libolm pickle instead of guessing
The unpickling from a libolm pickle was ignoring the `next_key_id` field
of the pickle, instead we looked at the key ids of the existing one-time
keys, picked the maximum value and incremented it.

The above method might work fine in most cases, but there isn't a
guarantee that we didn't already used up the one-time key with the
biggest key id.

The `next_key_id` was overlooked in the heat of developing vodozemac.